Add feature flag and new cards
